test_requirements.txt
# Test dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
coverage==7.3.2
hypothesis==6.92.0
